requirements.txt
python-dotenv==1.0.0
google-cloud-storage==2.7.0
